pyproject.toml
[project]
name = "rocket-booster"
version = "0.1.0"
description = "Make Browser Use browser agents faster with cached Playwright templates"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.34.0",
    "httpx>=0.28.0",
    "anthropic>=0.52.0",
    "browser-use>=0.2.0",
    "playwright>=1.52.0",
    "supabase>=2.15.0",
    "asyncpg>=0.30.0",
    "openai>=1.82.0",
    "python-dotenv>=1.1.0",
    "pydantic>=2.11.0",
    "langchain-anthropic>=0.3.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.25.0",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
markers = [
    "integration: marks tests that require real API keys (skip by default)",
]
